Added card matching behavior.
Now cards are correctly matched accordingly to the data model (still
mock up). After matching, the two cards have a thick on them and are
no more clickable. This is done through an div with class .overlay
(the style of which has been added to the CSS). Furthermore the board
now is inactive while performing checks and it supports end of the
game by matching all cards.